进行文件的逻辑拆分
初步对ui.py以及novel_generator.py进行了拆分
