+++ .gitignore
@@ -9,3 +9,5 @@
 /example
 config.json
 config_test.json
+/novel_generator/__pycache__
+/ui/__pycache__
