+++ novel_generator_backup.txt
