+++ ui/__init__.py
@@ -1,0 +1,2 @@
+# ui/__init__.py
+from .main_window import NovelGeneratorGUI
